requirements.txt
django==5.1.4
djangorestframework==3.14.0
mysqlclient==2.2.1
djoser==2.2.2
